v0.4.0 - Added members and modified names to CompressorStagePerformance
Able to track a lot more characteristics of the compressor.  This commit
comprises a lot of added and modified functionality over the course of 3-
4 months.  See files to see specific changes.  In summary this library
is used in Comp1dTools to extract compressor results when run at design
or analysis mode and can write out a bunch of compressor information.
It should be a really good start to creating a one-d meanline code!!!